Move resolveUrl to its own file
